requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.5.0
numpy>=1.24.0
diffusers>=0.17.0
pandas>=2.0.0
scipy>=1.10.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
pytest>=7.3.0
pytest-asyncio>=0.21.0
moviepy>=1.0.3
soundfile>=0.12.1
streamlit>=1.39.0
watchdog>=3.0.0
black
isort
mypy
flake8
